ip: don't use program name to select command
ip has an ancent behaviour of looking at its program name to determine
the command to run. If the name is longer than 2 characters, the first two
letters are stripped and the others are interpreted as the command name:

    $ ln -s /sbin/ip iproute
    $ ln -s /sbin/ip ipa
    $ ./iproute
    default via 192.168.55.1 dev wlp0s20f3 proto dhcp metric 600
    192.168.55.0/24 dev wlp0s20f3 proto kernel scope link src 192.168.55.26 metric 600
    192.168.122.0/24 dev virbr0 proto kernel scope link src 192.168.122.1 linkdown
    $ ./ipa show dev lo
    1: lo: <LOOPBACK,UP,LOWER_UP> mtu 65536 qdisc noqueue state UNKNOWN group default qlen 1000
        link/loopback 00:00:00:00:00:00 brd 00:00:00:00:00:00
        inet 127.0.0.1/8 scope host lo
           valid_lft forever preferred_lft forever

This creates problems when the ip binary is renamed. For example, Yocto
renames it to 'ip.iproute2' when also the busybox implementation is
present, giving the following error:

    $ ip.iproute2
    Object ".iproute2" is unknown, try "ip help".

Since noone is using it, remove this undocumented feature.

Signed-off-by: Matteo Croce <mcroce@microsoft.com>